Fix: FDISK not working on korean MSX computers
The problem (visible at least in the Daewoo CPC-400) is that
the CHPUT routine is supposed to not modify the IX and IY registers,
but in this computer it does since it does an inter-slot call to
an internal slot (presumably to handle korean characters); this causes
the PRINTF routine used by FDISK to crash.

Additionally, a small hack is added to print one less "-" character
in the upper and lower rulers when that inter-slot call is performed,
otherwise the FDISK screen isn't printed properly (it introduces an
additional line break for some reason).