Make the DOSVER function call return 6Fh in C in DOS 1 mode
The MSX-DOS 2 documentation says that DOSVER in DOS 1 returns B=0
and C undefined, but the fact is that in MSX-DOS 1 register C is
preserved (so 6Fh) and seems like some programs rely on that fact
to detect the DOS 1 mode. Nextor was corrupting C in this case,
so even if strictly speaking that shouldn't be an issue, this commit
changes the behavior of DOSVER in DOS 1 to return C=6Fh.