Fix DSKIO driver error check in DOS 1, page 1 transfers
The code in drv.mac was assuming that the call to the driver routine
DEV_RW/DSKIO was setting Cy=1 to signal an error, however this is true
only for drive-based drivers, but not for device-based drivers.
This incorrect behavior was happening only in DOS 1 mode and only
for transfers to page 1.